- New class created for storing music event flags and associated data. - Convenience class added for marking specific beats in a song, and converting between beats and seconds. - Tempo is now a float value. - States window has been removed, since it's unlikely an entirely separate dialogue box will be necessary. - Widgets that previously attached to ClipItem now attach instead to the more generic BaseMusicItem, allowing mixing of all music item types. - Mixer now holds a "snap to beat" option that changes which beat unit to snap to when editing clips.

+++ include/MusicPropertiesWidget.h
@@ -25,7 +25,7 @@ public:
 	double playbackSpeed();
 
 public slots:
-	void tempoChanged(int t);
+	void tempoChanged(double t);
 	void setTempo(int t);
 	void beatsPerMeasureChanged(int b);
 	void setBeatsPerMeasure(int b);
